streamlit_app/pyproject.toml
[project]
name = "nyc-real-estate-gpt"
version = "0.1.0"
description = ""
authors = [
    {name = "NIKHIL KUMAR REDDY",email = "nikhilkumar.m@skil.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "nycdb (>=0.3.4,<0.4.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.59.8,<2.0.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "instructor (>=1.7.2,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
